requirements.txt
flask
python-docx
vertexai
google-cloud-aiplatform